Fixed seek position to be Long rather than Int - causing overflow problems

+++ src/main/scala/com/esri/gdb/GDBIndexRow.scala
@@ -1,3 +1,3 @@
 package com.esri.gdb
 
-case class GDBIndexRow(oid: Int, seek: Int)
+case class GDBIndexRow(oid: Int, seek: Long)
